requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
art==5.7
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==4.1.7
djangorestframework==3.12.4
dnspython==2.3.0
email-validator==1.3.1
et-xmlfile==1.1.0
fastapi==0.95.0
flake8==6.0.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
IMAPClient==2.3.1
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
magic-filter==1.0.9
MarkupSafe==2.1.2
mccabe==0.7.0
multidict==6.0.4
numpy==1.23.4
openpyxl==3.1.2
orjson==3.8.8
Pillow==9.4.0
pluggy==0.13.1
prettytable==3.5.0
py==1.11.0
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
pyserial==3.5
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2021.1
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sorl-thumbnail==12.7.0
sqlparse==0.4.3
starlette==0.26.1
toml==0.10.2
typing_extensions==4.5.0
tzdata==2022.7
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
yarl==1.8.2
